Add React Router for navigation between views
- Install react-router-dom package
- Create separate page components for Employees and Offices views
- Implement routing configuration with nested routes
- Update App component to use Outlet for rendering routes
- Refactor navigation to use proper routing
