+++ packages/e2e.sandbox.peopleapp/package.json
@@ -44,6 +44,7 @@
     "core-js": "^3.37.1",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-router-dom": "^6.23.1",
     "swr": "^2.2.5",
     "tiny-invariant": "^1.3.3"
   },
