+++ packages/e2e.sandbox.peopleapp/src/main.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.js";
 import "./index.css";
 import { OsdkProvider2 } from "@osdk/react/experimental";
 import { $ } from "./foundryClient.js";
+import { Router } from "./router.js";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <OsdkProvider2 client={$}>
-      <App />
+      <Router />
     </OsdkProvider2>
   </React.StrictMode>,
 );
