+++ packages/e2e.sandbox.peopleapp/src/router.tsx
@@ -1,0 +1,29 @@
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import App from "./App.js";
+import { EmployeesPage } from "./pages/EmployeesPage.js";
+import { OfficesPage } from "./pages/OfficesPage.js";
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <App />,
+    children: [
+      {
+        path: "/",
+        element: <EmployeesPage />,
+      },
+      {
+        path: "/employees",
+        element: <EmployeesPage />,
+      },
+      {
+        path: "/offices",
+        element: <OfficesPage />,
+      },
+    ],
+  },
+]);
+
+export function Router() {
+  return <RouterProvider router={router} />;
+}
